blablabla

+++ commands/clear.js
@@ -5,9 +5,10 @@ module.exports = {
         if(!args[0] || isNaN(args[0]) || args[0] > 50 || args[0] < 1){
             return msg.reply("Põem uma porra decente ai, caralho!");
         }
-
+        
+        if(msg.author.name === "The_Joker"){
         await msg.channel.messages.fetch({limit: args[0]}).then(messages => {
             msg.channel.bulkDelete(messages);
-        })
+        })}
     }
 }
